- Added TileSet, and organisation for getting tiles from sets similar to namespaces.

+++ player_controller.h
@@ -20,8 +20,8 @@ namespace base {
 
 	protected:
 
-		float mMoveX;
-		float mMoveY;
+		int mMoveX;
+		int mMoveY;
 		bool mRunning;
 	};
